Showing email edit text only when event date is filled.
